modify インスタンスの生成はすべて factory で行うように変更
